requirements.txt
apache-beam[gcp]
google-api-python-client
google-cloud-datastore
